+++ sdk/ai/azure-ai-projects/azure/ai/projects/aio/operations/_operations.py
@@ -57,7 +57,7 @@ from ...operations._operations import (
     build_beta_evaluators_create_version_request,
     build_beta_evaluators_delete_version_request,
     build_beta_evaluators_get_version_request,
-    build_beta_evaluators_list_latest_versions_request,
+    build_beta_evaluators_list_request,
     build_beta_evaluators_list_versions_request,
     build_beta_evaluators_update_version_request,
     build_beta_insights_generate_request,
@@ -225,7 +225,7 @@ class AgentsOperations:
         *,
         name: str,
         definition: _models.AgentDefinition,
-        foundry_features: Optional[Union[str, _models.AgentDefinitionFeatureKeys]] = None,
+        foundry_features: Optional[Union[str, _models.AgentDefintionOptInKeys]] = None,
         content_type: str = "application/json",
         metadata: Optional[dict[str, str]] = None,
         description: Optional[str] = None,
@@ -236,7 +236,7 @@ class AgentsOperations:
         self,
         body: JSON,
         *,
-        foundry_features: Optional[Union[str, _models.AgentDefinitionFeatureKeys]] = None,
+        foundry_features: Optional[Union[str, _models.AgentDefintionOptInKeys]] = None,
         content_type: str = "application/json",
         **kwargs: Any
     ) -> _models.AgentDetails: ...
@@ -245,7 +245,7 @@ class AgentsOperations:
         self,
         body: IO[bytes],
         *,
-        foundry_features: Optional[Union[str, _models.AgentDefinitionFeatureKeys]] = None,
+        foundry_features: Optional[Union[str, _models.AgentDefintionOptInKeys]] = None,
         content_type: str = "application/json",
         **kwargs: Any
     ) -> _models.AgentDetails: ...
@@ -257,7 +257,7 @@ class AgentsOperations:
         *,
         name: str = _Unset,
         definition: _models.AgentDefinition = _Unset,
-        foundry_features: Optional[Union[str, _models.AgentDefinitionFeatureKeys]] = None,
+        foundry_features: Optional[Union[str, _models.AgentDefintionOptInKeys]] = None,
         metadata: Optional[dict[str, str]] = None,
         description: Optional[str] = None,
         **kwargs: Any
@@ -279,7 +279,7 @@ class AgentsOperations:
         :keyword foundry_features: A feature flag opt-in required when using preview operations or
          modifying persisted preview resources. Known values are: "HostedAgents=V1Preview" and
          "WorkflowAgents=V1Preview". Default value is None.
-        :paramtype foundry_features: str or ~azure.ai.projects.models.AgentDefinitionFeatureKeys
+        :paramtype foundry_features: str or ~azure.ai.projects.models.AgentDefintionOptInKeys
         :keyword metadata: Set of 16 key-value pairs that can be attached to an object. This can be
          useful for storing additional information about the object in a structured
          format, and querying for objects via API or the dashboard.
@@ -370,7 +370,7 @@ class AgentsOperations:
         agent_name: str,
         *,
         definition: _models.AgentDefinition,
-        foundry_features: Optional[Union[str, _models.AgentDefinitionFeatureKeys]] = None,
+        foundry_features: Optional[Union[str, _models.AgentDefintionOptInKeys]] = None,
         content_type: str = "application/json",
         metadata: Optional[dict[str, str]] = None,
         description: Optional[str] = None,
@@ -382,7 +382,7 @@ class AgentsOperations:
         agent_name: str,
         body: JSON,
         *,
-        foundry_features: Optional[Union[str, _models.AgentDefinitionFeatureKeys]] = None,
+        foundry_features: Optional[Union[str, _models.AgentDefintionOptInKeys]] = None,
         content_type: str = "application/json",
         **kwargs: Any
     ) -> _models.AgentDetails: ...
@@ -392,7 +392,7 @@ class AgentsOperations:
         agent_name: str,
         body: IO[bytes],
         *,
-        foundry_features: Optional[Union[str, _models.AgentDefinitionFeatureKeys]] = None,
+        foundry_features: Optional[Union[str, _models.AgentDefintionOptInKeys]] = None,
         content_type: str = "application/json",
         **kwargs: Any
     ) -> _models.AgentDetails: ...
@@ -404,7 +404,7 @@ class AgentsOperations:
         body: Union[JSON, IO[bytes]] = _Unset,
         *,
         definition: _models.AgentDefinition = _Unset,
-        foundry_features: Optional[Union[str, _models.AgentDefinitionFeatureKeys]] = None,
+        foundry_features: Optional[Union[str, _models.AgentDefintionOptInKeys]] = None,
         metadata: Optional[dict[str, str]] = None,
         description: Optional[str] = None,
         **kwargs: Any
@@ -422,7 +422,7 @@ class AgentsOperations:
         :keyword foundry_features: A feature flag opt-in required when using preview operations or
          modifying persisted preview resources. Known values are: "HostedAgents=V1Preview" and
          "WorkflowAgents=V1Preview". Default value is None.
-        :paramtype foundry_features: str or ~azure.ai.projects.models.AgentDefinitionFeatureKeys
+        :paramtype foundry_features: str or ~azure.ai.projects.models.AgentDefintionOptInKeys
         :keyword metadata: Set of 16 key-value pairs that can be attached to an object. This can be
          useful for storing additional information about the object in a structured
          format, and querying for objects via API or the dashboard.
@@ -942,7 +942,7 @@ class AgentsOperations:
         agent_name: str,
         *,
         definition: _models.AgentDefinition,
-        foundry_features: Optional[Union[str, _models.AgentDefinitionFeatureKeys]] = None,
+        foundry_features: Optional[Union[str, _models.AgentDefintionOptInKeys]] = None,
         content_type: str = "application/json",
         metadata: Optional[dict[str, str]] = None,
         description: Optional[str] = None,
@@ -963,7 +963,7 @@ class AgentsOperations:
         :keyword foundry_features: A feature flag opt-in required when using preview operations or
          modifying persisted preview resources. Known values are: "HostedAgents=V1Preview" and
          "WorkflowAgents=V1Preview". Default value is None.
-        :paramtype foundry_features: str or ~azure.ai.projects.models.AgentDefinitionFeatureKeys
+        :paramtype foundry_features: str or ~azure.ai.projects.models.AgentDefintionOptInKeys
         :keyword content_type: Body Parameter content-type. Content type parameter for JSON body.
          Default value is "application/json".
         :paramtype content_type: str
@@ -987,7 +987,7 @@ class AgentsOperations:
         agent_name: str,
         body: JSON,
         *,
-        foundry_features: Optional[Union[str, _models.AgentDefinitionFeatureKeys]] = None,
+        foundry_features: Optional[Union[str, _models.AgentDefintionOptInKeys]] = None,
         content_type: str = "application/json",
         **kwargs: Any
     ) -> _models.AgentVersionDetails:
@@ -1005,7 +1005,7 @@ class AgentsOperations:
         :keyword foundry_features: A feature flag opt-in required when using preview operations or
          modifying persisted preview resources. Known values are: "HostedAgents=V1Preview" and
          "WorkflowAgents=V1Preview". Default value is None.
-        :paramtype foundry_features: str or ~azure.ai.projects.models.AgentDefinitionFeatureKeys
+        :paramtype foundry_features: str or ~azure.ai.projects.models.AgentDefintionOptInKeys
         :keyword content_type: Body Parameter content-type. Content type parameter for JSON body.
          Default value is "application/json".
         :paramtype content_type: str
@@ -1020,7 +1020,7 @@ class AgentsOperations:
         agent_name: str,
         body: IO[bytes],
         *,
-        foundry_features: Optional[Union[str, _models.AgentDefinitionFeatureKeys]] = None,
+        foundry_features: Optional[Union[str, _models.AgentDefintionOptInKeys]] = None,
         content_type: str = "application/json",
         **kwargs: Any
     ) -> _models.AgentVersionDetails:
@@ -1038,7 +1038,7 @@ class AgentsOperations:
         :keyword foundry_features: A feature flag opt-in required when using preview operations or
          modifying persisted preview resources. Known values are: "HostedAgents=V1Preview" and
          "WorkflowAgents=V1Preview". Default value is None.
-        :paramtype foundry_features: str or ~azure.ai.projects.models.AgentDefinitionFeatureKeys
+        :paramtype foundry_features: str or ~azure.ai.projects.models.AgentDefintionOptInKeys
         :keyword content_type: Body Parameter content-type. Content type parameter for binary body.
          Default value is "application/json".
         :paramtype content_type: str
@@ -1054,7 +1054,7 @@ class AgentsOperations:
         body: Union[JSON, IO[bytes]] = _Unset,
         *,
         definition: _models.AgentDefinition = _Unset,
-        foundry_features: Optional[Union[str, _models.AgentDefinitionFeatureKeys]] = None,
+        foundry_features: Optional[Union[str, _models.AgentDefintionOptInKeys]] = None,
         metadata: Optional[dict[str, str]] = None,
         description: Optional[str] = None,
         **kwargs: Any
@@ -1076,7 +1076,7 @@ class AgentsOperations:
         :keyword foundry_features: A feature flag opt-in required when using preview operations or
          modifying persisted preview resources. Known values are: "HostedAgents=V1Preview" and
          "WorkflowAgents=V1Preview". Default value is None.
-        :paramtype foundry_features: str or ~azure.ai.projects.models.AgentDefinitionFeatureKeys
+        :paramtype foundry_features: str or ~azure.ai.projects.models.AgentDefintionOptInKeys
         :keyword metadata: Set of 16 key-value pairs that can be attached to an object. This can be
          useful for storing additional information about the object in a structured
          format, and querying for objects via API or the dashboard.
@@ -4168,7 +4168,7 @@ class BetaEvaluatorsOperations:
         return AsyncItemPaged(get_next, extract_data)
 
     @distributed_trace
-    def list_latest_versions(
+    def list(
         self,
         *,
         type: Optional[Union[Literal["builtin"], Literal["custom"], Literal["all"], str]] = None,
@@ -4207,7 +4207,7 @@ class BetaEvaluatorsOperations:
         def prepare_request(next_link=None):
             if not next_link:
 
-                _request = build_beta_evaluators_list_latest_versions_request(
+                _request = build_beta_evaluators_list_request(
                     foundry_features=foundry_features,
                     type=type,
                     limit=limit,
@@ -4835,11 +4835,13 @@ class BetaInsightsOperations:
         return deserialized  # type: ignore
 
     @distributed_trace_async
-    async def get(self, id: str, *, include_coordinates: Optional[bool] = None, **kwargs: Any) -> _models.Insight:
+    async def get(
+        self, insight_id: str, *, include_coordinates: Optional[bool] = None, **kwargs: Any
+    ) -> _models.Insight:
         """Get a specific insight by Id.
 
-        :param id: The unique identifier for the insights report. Required.
-        :type id: str
+        :param insight_id: The unique identifier for the insights report. Required.
+        :type insight_id: str
         :keyword include_coordinates: Whether to include coordinates for visualization in the response.
          Defaults to false. Default value is None.
         :paramtype include_coordinates: bool
@@ -4864,7 +4866,7 @@ class BetaInsightsOperations:
         cls: ClsType[_models.Insight] = kwargs.pop("cls", None)
 
         _request = build_beta_insights_get_request(
-            id=id,
+            insight_id=insight_id,
             foundry_features=foundry_features,
             include_coordinates=include_coordinates,
             api_version=self._config.api_version,
@@ -6389,11 +6391,11 @@ class BetaSchedulesOperations:
         self._deserialize: Deserializer = input_args.pop(0) if input_args else kwargs.pop("deserializer")
 
     @distributed_trace_async
-    async def delete(self, id: str, **kwargs: Any) -> None:
+    async def delete(self, schedule_id: str, **kwargs: Any) -> None:
         """Delete a schedule.
 
-        :param id: Identifier of the schedule. Required.
-        :type id: str
+        :param schedule_id: Identifier of the schedule. Required.
+        :type schedule_id: str
         :return: None
         :rtype: None
         :raises ~azure.core.exceptions.HttpResponseError:
@@ -6415,7 +6417,7 @@ class BetaSchedulesOperations:
         cls: ClsType[None] = kwargs.pop("cls", None)
 
         _request = build_beta_schedules_delete_request(
-            id=id,
+            schedule_id=schedule_id,
             foundry_features=foundry_features,
             api_version=self._config.api_version,
             headers=_headers,
@@ -6441,11 +6443,11 @@ class BetaSchedulesOperations:
             return cls(pipeline_response, None, {})  # type: ignore
 
     @distributed_trace_async
-    async def get(self, id: str, **kwargs: Any) -> _models.Schedule:
+    async def get(self, schedule_id: str, **kwargs: Any) -> _models.Schedule:
         """Get a schedule by id.
 
-        :param id: Identifier of the schedule. Required.
-        :type id: str
+        :param schedule_id: Identifier of the schedule. Required.
+        :type schedule_id: str
         :return: Schedule. The Schedule is compatible with MutableMapping
         :rtype: ~azure.ai.projects.models.Schedule
         :raises ~azure.core.exceptions.HttpResponseError:
@@ -6467,7 +6469,7 @@ class BetaSchedulesOperations:
         cls: ClsType[_models.Schedule] = kwargs.pop("cls", None)
 
         _request = build_beta_schedules_get_request(
-            id=id,
+            schedule_id=schedule_id,
             foundry_features=foundry_features,
             api_version=self._config.api_version,
             headers=_headers,
@@ -6608,12 +6610,12 @@ class BetaSchedulesOperations:
 
     @overload
     async def create_or_update(
-        self, id: str, schedule: _models.Schedule, *, content_type: str = "application/json", **kwargs: Any
+        self, schedule_id: str, schedule: _models.Schedule, *, content_type: str = "application/json", **kwargs: Any
     ) -> _models.Schedule:
         """Create or update operation template.
 
-        :param id: Identifier of the schedule. Required.
-        :type id: str
+        :param schedule_id: Identifier of the schedule. Required.
+        :type schedule_id: str
         :param schedule: The resource instance. Required.
         :type schedule: ~azure.ai.projects.models.Schedule
         :keyword content_type: Body Parameter content-type. Content type parameter for JSON body.
@@ -6626,12 +6628,12 @@ class BetaSchedulesOperations:
 
     @overload
     async def create_or_update(
-        self, id: str, schedule: JSON, *, content_type: str = "application/json", **kwargs: Any
+        self, schedule_id: str, schedule: JSON, *, content_type: str = "application/json", **kwargs: Any
     ) -> _models.Schedule:
         """Create or update operation template.
 
-        :param id: Identifier of the schedule. Required.
-        :type id: str
+        :param schedule_id: Identifier of the schedule. Required.
+        :type schedule_id: str
         :param schedule: The resource instance. Required.
         :type schedule: JSON
         :keyword content_type: Body Parameter content-type. Content type parameter for JSON body.
@@ -6644,12 +6646,12 @@ class BetaSchedulesOperations:
 
     @overload
     async def create_or_update(
-        self, id: str, schedule: IO[bytes], *, content_type: str = "application/json", **kwargs: Any
+        self, schedule_id: str, schedule: IO[bytes], *, content_type: str = "application/json", **kwargs: Any
     ) -> _models.Schedule:
         """Create or update operation template.
 
-        :param id: Identifier of the schedule. Required.
-        :type id: str
+        :param schedule_id: Identifier of the schedule. Required.
+        :type schedule_id: str
         :param schedule: The resource instance. Required.
         :type schedule: IO[bytes]
         :keyword content_type: Body Parameter content-type. Content type parameter for binary body.
@@ -6662,12 +6664,12 @@ class BetaSchedulesOperations:
 
     @distributed_trace_async
     async def create_or_update(
-        self, id: str, schedule: Union[_models.Schedule, JSON, IO[bytes]], **kwargs: Any
+        self, schedule_id: str, schedule: Union[_models.Schedule, JSON, IO[bytes]], **kwargs: Any
     ) -> _models.Schedule:
         """Create or update operation template.
 
-        :param id: Identifier of the schedule. Required.
-        :type id: str
+        :param schedule_id: Identifier of the schedule. Required.
+        :type schedule_id: str
         :param schedule: The resource instance. Is one of the following types: Schedule, JSON,
          IO[bytes] Required.
         :type schedule: ~azure.ai.projects.models.Schedule or JSON or IO[bytes]
@@ -6700,7 +6702,7 @@ class BetaSchedulesOperations:
             _content = json.dumps(schedule, cls=SdkJSONEncoder, exclude_readonly=True)  # type: ignore
 
         _request = build_beta_schedules_create_or_update_request(
-            id=id,
+            schedule_id=schedule_id,
             foundry_features=foundry_features,
             content_type=content_type,
             api_version=self._config.api_version,
@@ -6813,7 +6815,7 @@ class BetaSchedulesOperations:
     @distributed_trace
     def list_runs(
         self,
-        id: str,
+        schedule_id: str,
         *,
         type: Optional[Union[str, _models.ScheduleTaskType]] = None,
         enabled: Optional[bool] = None,
@@ -6821,8 +6823,8 @@ class BetaSchedulesOperations:
     ) -> AsyncItemPaged["_models.ScheduleRun"]:
         """List all schedule runs.
 
-        :param id: Identifier of the schedule. Required.
-        :type id: str
+        :param schedule_id: Identifier of the schedule. Required.
+        :type schedule_id: str
         :keyword type: Filter by the type of schedule. Known values are: "Evaluation" and "Insight".
          Default value is None.
         :paramtype type: str or ~azure.ai.projects.models.ScheduleTaskType
@@ -6852,7 +6854,7 @@ class BetaSchedulesOperations:
             if not next_link:
 
                 _request = build_beta_schedules_list_runs_request(
-                    id=id,
+                    schedule_id=schedule_id,
                     foundry_features=foundry_features,
                     type=type,
                     enabled=enabled,