+++ sdk/ai/azure-ai-projects/azure/ai/projects/models/_enums.py
@@ -10,10 +10,8 @@ from enum import Enum
 from azure.core import CaseInsensitiveEnumMeta
 
 
-class AgentDefinitionFeatureKeys(str, Enum, metaclass=CaseInsensitiveEnumMeta):
-    """Feature opt-in keys for agent definition operations supporting container, hosted, and workflow
-    agents.
-    """
+class AgentDefintionOptInKeys(str, Enum, metaclass=CaseInsensitiveEnumMeta):
+    """Feature opt-in keys for agent definition operations supporting hosted or workflow agents."""
 
     HOSTED_AGENTS_V1_PREVIEW = "HostedAgents=V1Preview"
     """HOSTED_AGENTS_V1_PREVIEW."""
@@ -363,10 +361,6 @@ class EvaluatorType(str, Enum, metaclass=CaseInsensitiveEnumMeta):
 class FoundryFeaturesOptInKeys(str, Enum, metaclass=CaseInsensitiveEnumMeta):
     """Type of FoundryFeaturesOptInKeys."""
 
-    HOSTED_AGENTS_V1_PREVIEW = "HostedAgents=V1Preview"
-    """HOSTED_AGENTS_V1_PREVIEW."""
-    WORKFLOW_AGENTS_V1_PREVIEW = "WorkflowAgents=V1Preview"
-    """WORKFLOW_AGENTS_V1_PREVIEW."""
     EVALUATIONS_V1_PREVIEW = "Evaluations=V1Preview"
     """EVALUATIONS_V1_PREVIEW."""
     SCHEDULES_V1_PREVIEW = "Schedules=V1Preview"
@@ -552,8 +546,8 @@ class RankerVersionType(str, Enum, metaclass=CaseInsensitiveEnumMeta):
 
     AUTO = "auto"
     """AUTO."""
-    DEFAULT2024_11_15 = "default-2024-11-15"
-    """DEFAULT2024_11_15."""
+    DEFAULT_2024_11_15 = "default-2024-11-15"
+    """DEFAULT_2024_11_15."""
 
 
 class RecurrenceType(str, Enum, metaclass=CaseInsensitiveEnumMeta):
@@ -668,8 +662,8 @@ class ToolChoiceParamType(str, Enum, metaclass=CaseInsensitiveEnumMeta):
     """WEB_SEARCH_PREVIEW."""
     COMPUTER_USE_PREVIEW = "computer_use_preview"
     """COMPUTER_USE_PREVIEW."""
-    WEB_SEARCH_PREVIEW2025_03_11 = "web_search_preview_2025_03_11"
-    """WEB_SEARCH_PREVIEW2025_03_11."""
+    WEB_SEARCH_PREVIEW_2025_03_11 = "web_search_preview_2025_03_11"
+    """WEB_SEARCH_PREVIEW_2025_03_11."""
     IMAGE_GENERATION = "image_generation"
     """IMAGE_GENERATION."""
     CODE_INTERPRETER = "code_interpreter"