+++ sdk/ai/azure-ai-projects/samples/evaluations/sample_eval_catalog.py
@@ -177,13 +177,13 @@ with (
     )
 
     print("Getting list of builtin evaluator versions")
-    evaluators = project_client.beta.evaluators.list_latest_versions(type="builtin")
+    evaluators = project_client.beta.evaluators.list(type="builtin")
     print("List of builtin evaluator versions")
     for evaluator in evaluators:
         pprint(evaluator)
 
     print("Getting list of custom evaluator versions")
-    evaluators = project_client.beta.evaluators.list_latest_versions(type="custom")
+    evaluators = project_client.beta.evaluators.list(type="custom")
     print("List of custom evaluator versions")
     for evaluator in evaluators:
         pprint(evaluator)