+++ sdk/ai/azure-ai-projects/samples/evaluations/sample_evaluation_compare_insight.py
@@ -142,10 +142,10 @@ with (
                 ),
             ),
         )
-        print(f"Started insight generation (id: {compareInsight.id})")
+        print(f"Started insight generation (id: {compareInsight.insight_id})")
 
         while compareInsight.state not in [OperationState.SUCCEEDED, OperationState.FAILED]:
-            compareInsight = project_client.beta.insights.get(id=compareInsight.id)
+            compareInsight = project_client.beta.insights.get(insight_id=compareInsight.insight_id)
             print(f"Waiting for insight to be generated...current status: {compareInsight.state}")
             time.sleep(5)